backend/requirements.txt
#django settings
Django==4.0.7
djangorestframework==3.13.1
django-environ==0.9.0
djangorestframework-simplejwt==5.2.0

#mysqldb
pymysql==1.0.2
mysqlclient
cryptography

# swagger
drf-yasg==1.21.3

#s3 bucket
boto3==1.24.67
django-storages==1.13.1

python-dateutil==2.8.2
